@@ -18,13 +18,13 @@
     "daisyui": "^2.13.6",
     "faker": "^5.5.3",
     "flowbite": "^1.4.1",
-    "next": "12.1.4",
+    "next": "^13.0.5",
     "next-auth": "^4.3.1",
     "next-themes": "^0.1.1",
-    "react": "^18.0.0",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.6.0",
     "react-chartjs-2": "^4.1.0",
-    "react-dom": "^18.0.0",
+    "react-dom": "^18.2.0",
     "tailwindcss-border-gradients": "^2.0.3",
     "tailwindcss-padded-radius": "^1.0.0",
     "tailwindcss-radix": "^1.6.0",
@@ -40,7 +40,7 @@
     "@types/react": "^18.0.25",
     "autoprefixer": "^10.4.4",
     "eslint": "8.12.0",
-    "eslint-config-next": "12.1.4",
+    "eslint-config-next": "^13.0.5",
     "postcss": "^8.4.12",
     "postcss-import": "^14.1.0",
     "prettier": "^2.6.2",
